🚚 [move] FT-50 채팅 서비스 폴더 생성
