+++ srcs/back/src/modules/chat/service/chat.service.ts
@@ -1,14 +1,14 @@
 import { Injectable } from '@nestjs/common';
 import { compare as comparePassword } from 'bcryptjs';
-import { UsersService } from '../users/users.service';
+import { UsersService } from '../../users/users.service';
 import { ChannelService } from './channel.service';
 import { DirectMessageService } from './direct-message.service';
-import { CreateChannelDto } from './dto/create-channel.dto';
-import { CreateDirectMessageDto } from './dto/create-direct-message.dto';
-import { CreateMessageDto } from './dto/create-message.dto';
-import { UpdateChannelDto } from './dto/update-channel.dto';
-import { Channel } from './entity/channel.entity';
-import { DirectMessage } from './entity/direct-message.entity';
+import { CreateChannelDto } from '../dto/create-channel.dto';
+import { CreateDirectMessageDto } from '../dto/create-direct-message.dto';
+import { CreateMessageDto } from '../dto/create-message.dto';
+import { UpdateChannelDto } from '../dto/update-channel.dto';
+import { Channel } from '../entity/channel.entity';
+import { DirectMessage } from '../entity/direct-message.entity';
 import { MessageService } from './message.service';
 
 @Injectable()
